+++ packages/wafer-client-js/src/types/index.ts
@@ -1,5 +1,7 @@
 export type { WaferConfig } from './config';
 export type { WaferMessage, WaferMeta } from './message';
 export type { WaferResult, WaferResponse } from './result';
-export { WaferError } from './error';
-export type { WaferErrorCode } from './error';
+export { WaferError, isWaferServerErrorCode } from './error';
+export type { WaferErrorCode, WaferClientErrorCode } from './error';
+export { WAFER_SERVER_ERROR_CODES } from './error-codes';
+export type { WaferServerErrorCode } from './error-codes';